updated banner component
